+++ packages/stripe-plugin/src/index.ts
@@ -1,1 +1,3 @@
 export { StripePlugin } from './stripe.plugin';
+export type { StripePluginOptions } from './types';
+export type { StripeHttpClient, StripeLatestApiVersion } from './stripe-types';